+++ NEWS.md
@@ -1,10 +1,13 @@
 # Version 0.3.0
 
 ## Change in 0.3.0
-- Addition of `filterChromData()` and `filterPeaksData()` methods.
+- Addition of `filterChromData()` method for `ChromBackend`.
+- Creation of the `Chromatograms` class and implementation of basic accessor 
+  methods.
+- Addition of basic plotting functions. 
 
 ## Changes in 0.2.0
-- Addition of ChomBackendMemory class and associated methods 
+- Addition of `ChomBackendMemory` class and associated methods 
 
 ## Changes in 0.1.0
-- Addition of basic ChromBackend class and default methods 
+- Addition of basic `ChromBackend` class and default methods